+++ src/main/java/AttendanceApplication.java
@@ -1,4 +1,4 @@
-import presentation.AttendanceController;
+import controller.AttendanceController;
 import view.AttendanceFileReader;
 import view.InputView;
 import view.OutputView;
