+++ src/main/java/controller/command/CrewInfoCommand.java
@@ -1,9 +1,9 @@
-package presentation.command;
+package controller.command;
 
 import domain.AttendanceBook;
 import java.util.function.Consumer;
 
-public class EditCommand implements Consumer<AttendanceBook> {
+public class CrewInfoCommand implements Consumer<AttendanceBook> {
     @Override
     public void accept(AttendanceBook attendanceBook) {
