+++ src/main/java/controller/command/WarningInfoCommand.java
@@ -1,9 +1,9 @@
-package presentation.command;
+package controller.command;
 
 import domain.AttendanceBook;
 import java.util.function.Consumer;
 
-public class CrewInfoCommand implements Consumer<AttendanceBook> {
+public class WarningInfoCommand implements Consumer<AttendanceBook> {
     @Override
     public void accept(AttendanceBook attendanceBook) {
